backend/requirements.txt
fastapi==0.115.12
firecrawl==1.15.0
python-jose==3.4.0
langchain==0.3.23
passlib==1.7.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic-core==2.33.1
email-validator==2.2.0
SQLAlchemy==2.0.30
psycopg2==2.9.10
uvicorn==0.34.0
python-multipart==0.0.20
bcrypt==4.3.0